fix(pa): share constraint-value normalizer and add defensive guards
Prerequisite refactor split out of maggie-li-yd/2872-webmcp-poc (PR A1). No
functional change to existing UI flows.

- Extract value normalization out of useDashboardFlow into
  utils/applyConstraintValue.ts so the wizard flow and the (upcoming) WebMCP
  patch applier share one normalizer. Reject objects that are neither a
  {from,to} date range nor a numeric value instead of String()-ing them into
  a broken '[object Object]' filter.
- StackBarChart: null-guard filterCard?.name so a category referencing a
  no-longer-chartable filter card can't crash the whole chart render.
- BaseTagInput: coerce displayValue to String() so a numeric constraint value
  doesn't throw in .replace().
